requirements.txt
aiohttp==3.9.1
fastapi==0.103.1
uvicorn==0.23.2
redis==4.6.0
pydantic==2.3.0
langchain==0.0.267
langchain-core==0.1.16
langchain-openai==0.0.2
python-dotenv==1.0.0
pytest==7.4.2
httpx==0.24.1
tenacity==8.2.3 